requirements.txt
Flask==2.0.3
Flask-CORS==3.0.10
gTTS==2.2.3
fpdf2==2.5.5
pymongo==3.12.1
deep-translator==1.8.1
speech_recognition==3.8.1
indic-transliteration==1.1.0
PyPDF2==1.26.0
python-dotenv==0.19.2
gunicorn==20.1.0